Split the billing plans grid into account-billing-plans.tsx
account-billing.tsx crossed the 800-line file-size ratchet. The tier
cards (plan definitions, rank helpers, prorated switch labels, interval
radios, and subscribe/switch buttons) move to a sibling render helper;
the route keeps loading, checkout, status, and cancellation feedback.
Behaviour is unchanged.
